Move exact printing to submodule and introduce dummy top level module
