+++ tests/Test.hs
@@ -2,10 +2,10 @@
 -- | Use "runhaskell Setup.hs test" or "cabal test" to run these tests.
 module Main where
 
-import Language.Haskell.GHC.ExactPrint  ()
+import Language.Haskell.GHC.ExactPrint
+import Language.Haskell.GHC.ExactPrint.Utils
 import Language.Haskell.GHC.ExactPrint.Types
-import Language.Haskell.GHC.ExactPrint.Utils (ss2span, debug, showGhc, showAnnData, ghead)
-import Language.Haskell.GHC.ExactPrint.Common
+
 
 import GHC.Paths ( libdir )
